refactor: extract SyncButton from MainFrame components
