+++ src/stores/github.store.ts
@@ -190,7 +190,7 @@ export const fileSysyemStore = create<fileSysyemPropsStore>()((set, get) => ({
   otherUserSelectedCommitFile: "",
   commitFileList: [],
   commentsList: [],
-  clickedFileList: [],
+  clickedFileList: [ADDITIONAL_FILES.DEFAULT_FILE],
   setSelectedCommitFile: (newFile) => {
     set((state) => {
       const isFileInList = state.clickedFileList.some(
